requirements.txt
# Install basic packages
openai 
anthropic 
python-dotenv 
numpy 
pandas 
matplotlib 
seaborn 
jupyter